Add `!` versions of save and update (#82)
* Add save! method on LuckyRecord::Form

This method will work like the normal save, but only take parameters,
without a block. If it saves successfully, it will return the record,
_not_ the form. If it fails, it will raise a
LuckyRecord::Exceptions::InvalidForm error.

I tried to use the existing logic where it made sense, since the goal is
for the method to work the same but have some different behavior at the
end of the save.

* Add update! on LuckyRecord::Form

This method will work like the normal update, but only take parameters
and the record, without a block. If it updates successfully, it will
return the record, _not_ the form. If it fails, it will raise a
LuckyRecord::Exceptions::InvalidForm error.

I tried to use the existing logic where it made sense, since the goal is
for the method to work the same but have some different behavior at the
end of the update. In this case, it calls save! because it has all the
logic and return values already in place.

* Add more specific InvalidFormError

This error object now takes in more information to display a better
error message for the user.