Fix preload bug when loading results twice
Closes #51

Here is the problem. We were add the `in(ids)` every time a preload proc
was called. So the second time `results` was called, it would add
another identical clause. This was a no-no.

Instead we duplicate the query so we can get a fresh copy of the query
that we use just for preloading w/o modifying the original query.