Refactor adding columns
Previously to add a column we had to add a bunch of methods and method
overloads to multiple modules. This made things hard to follow and
error-prone.

Now we have classes that handle adding columns. It makes it easy to add
new columns and even third parties can add new columns by making a new
class. I think there are still some changes we can make to make this
easier, but this is a step in a good direction IMO.