Table name is guessed. No requires required (#132)
This improves an issue where previously you had to require associations.
Now we don't because we no longer resolve the type in the macro.

The one downside is that if you have a table name that is not the
default pluralized snake case, you will need to provide it in
`belongs_to`. I think this is a great tradeoff since the require issue
is annoying and causes a lot of issues. Also most people use those
defaults so it should be fine.