Schema enforcer update (#556)
In order for views to work with SchemaEnforcer, EnsureMatchingColumns has to have a configuration to not check required vs. nilable on columns since all columns are nilable in views. This update also changes how we set up validations for SchemaEnforcer so that tables could have different validations than views and there's a way to add/remove/customize validations