Add extract functionality as a new criteria to timestamptz columns
This PR concerns the extract info from date mentioned in issue #857.
It provides a way to be able to extract specific fields from a timestamptz column and
compare it with some numerical values.
For all the fields implemented till now (see Avram::ChronoUnits in the extract_criteria),
three possible methods are provided:

- column.extract(Avram::ChronoUnits::<enum_value>)     e.g: created_on.extract(Avram::ChronoUnits::Day)
- column.extract(<enum_value_as_symbol>)               e.g: created_on.extract(:day)
- column.extract_<enum_value>                          e.g: created_on.extract_day

Most values returned are ints, but 4 of them which return floats
- julian, epoch, second, milliseconds

Functionality can be used as:
- column.extract_year.eq(2022)