Introduce error message when parsing invalid values.
Implements #893. There is a breaking change, though.
Previously, calling FailedCast#value would always return a nil value, without providing any clue on which value
was actually breaking.
Now, it raises a FailCastError with a talking error message, in the form of - for instance - "Value ->Invalid<- is
not a valid constant for enum WhateverEnum".